requirements.txt
# Core científico
numpy>=1.24.0
pandas>=2.0.0

# Modelos
statsmodels>=0.14.0
chronos-forecasting>=1.0.0

# Configuración
pyyaml>=6.0
